fix: remove unnecessary u64 cast and unused import for CI
- Remove redundant u64 cast in stream_source_service.rs
- Remove unused Duration import in job_scheduler tests
- Both issues were causing CI failures with clippy warnings

+++ crates/m3u-proxy/src/services/stream_source_service.rs
@@ -373,8 +373,7 @@ impl StreamSourceService {
             .stream_source_repo
             .get_channel_count_for_source(&id)
             .await
-            .map_err(|e| anyhow::anyhow!("Failed to get channel count: {}", e))?
-            as u64;
+            .map_err(|e| anyhow::anyhow!("Failed to get channel count: {}", e))?;
 
         // Use service pattern for URL linking
         let linked_epgs = self
